# services/kloud_bridge/openapi/kloud-bridge-openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clisonix Kloud Bridge",
    "description": "Isolated bridge microservice that connects Clisonix to the sovereign Kloud fabric without merging codebases.",
    "version": "0.5.0",
    "x-clisonix-service": "kloud-bridge",
    "x-exported-at": "2026-04-04T13:55:23.066752+00:00"
  },
  "paths": {
...
    "/api/v1/fabric/sync": {
      "post": {
        "summary": "Fabric Sync",
        "operationId": "fabric_sync_api_v1_fabric_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FabricSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Fabric Sync Api V1 Fabric Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fabric/sync": {
      "post": {
        "summary": "Fabric Sync",
        "operationId": "fabric_sync_fabric_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FabricSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Fabric Sync Fabric Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FabricSyncRequest": {
        "properties": {
          "include_state": {
            "type": "boolean",
            "title": "Include State",
            "default": true
          },
          "include_peers": {
            "type": "boolean",
            "title": "Include Peers",
            "default": true
          },
          "include_status": {
            "type": "boolean",
            "title": "Include Status",
            "default": true
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "FabricSyncRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}